move the Quaternion class from quantum_info/synthesis to quantum_info
