+++ qiskit/quantum_info/quaternion.py
