+++ qiskit/quantum_info/synthesis/__init__.py
@@ -15,5 +15,4 @@
 from __future__ import annotations
 from .two_qubit_decompose import TwoQubitBasisDecomposer, two_qubit_cnot_decompose
 from .one_qubit_decompose import OneQubitEulerDecomposer
-from .quaternion import Quaternion
 from .xx_decompose.decomposer import XXDecomposer
